+++ lib/honeybadger/breadcrumbs/active_support.rb
@@ -15,17 +15,19 @@ module Honeybadger
               ["Active Record", name].compact.join(" - ")
             end,
             category: "query",
-            select_keys: [:sql, :name, :connection_id, :cached],
+            select_keys: [:sql, :name, :connection, :connection_id, :cached],
             transform: lambda do |data|
               if data[:sql]
-                adapter = ::ActiveRecord::Base.connection_config[:adapter]
+                connection = data.delete(:connection)
+                adapter = (connection && connection.adapter_name.downcase) || active_record_connection_db_config[:adapter]
                 data[:sql] = Util::SQL.obfuscate(data[:sql], adapter)
               end
               data
             end,
             exclude_when: lambda do |data|
               # Ignore schema, begin, and commit transaction queries
-              data[:name] == "SCHEMA" || (data[:sql] && (data[:sql] =~ /^(begin|commit)( transaction)?$/i))
+              data[:name] == "SCHEMA" ||
+                (data[:sql] && (Util::SQL.force_utf_8(data[:sql].dup) =~ /^(begin|commit)( transaction)?$/i))
             end
           },
 
@@ -103,6 +105,15 @@ module Honeybadger
         }
       end
 
+      private_class_method def self.active_record_connection_db_config
+        if ::ActiveRecord::Base.respond_to?(:connection_db_config)
+          # >= Rails 6.1
+          ::ActiveRecord::Base.connection_db_config.configuration_hash
+        else
+          # < Rails 6.1
+          ::ActiveRecord::Base.connection_config
+        end
+      end
     end
   end
 end