+++ lib/honeybadger/plugins/rails.rb
@@ -29,6 +29,17 @@ module Honeybadger
         end
       end
 
+      class ErrorSubscriber
+        def self.report(exception, handled:, severity:, context: {}, source: nil)
+          return if source && ::Honeybadger.config[:'rails.subscriber_ignore_sources'].any? { |regex| regex.match?(source) }
+
+          tags = ["severity:#{severity}", "handled:#{handled}"]
+          tags << "source:#{source}" if source
+          Honeybadger.notify(exception, context: context, tags: tags)
+          exception.instance_variable_set(:@__hb_handled, true)
+        end
+      end
+
       Plugin.register :rails_exceptions_catcher do
         requirement { defined?(::Rails.application) && ::Rails.application }
 
@@ -40,6 +51,11 @@ module Honeybadger
           elsif defined?(::ActionDispatch::ShowExceptions)
             # Rails 3.0.x and 3.1.x
             ::ActionDispatch::ShowExceptions.prepend(ExceptionsCatcher)
+          end
+
+          if defined?(::ActiveSupport::ErrorReporter)
+            # Rails 7
+            ::Rails.error.subscribe(ErrorSubscriber)
           end
         end
       end