+++ lib/honeybadger/util/revision.rb
@@ -3,9 +3,14 @@ module Honeybadger
     class Revision
       class << self
         def detect(root = Dir.pwd)
-          from_heroku ||
+          revision = from_heroku ||
             from_capistrano(root) ||
             from_git
+
+          revision = revision.to_s.strip
+          return unless revision =~ /\S/
+
+          revision
         end
 
         private
@@ -21,12 +26,12 @@ module Honeybadger
         def from_capistrano(root)
           file = File.join(root, 'REVISION')
           return nil unless File.file?(file)
-          File.read(file).strip rescue nil
+          File.read(file) rescue nil
         end
 
         def from_git
           return nil unless File.directory?('.git')
-          `git rev-parse HEAD`.strip rescue nil
+          `git rev-parse HEAD 2> #{File::NULL}` rescue nil
         end
       end
     end