+++ spec/unit/honeybadger/notice_spec.rb
@@ -64,6 +64,20 @@ describe Honeybadger::Notice do
     expect(build_notice(controller: 'users_controller').controller).to eq 'users_controller'
   end
 
+  it "aliases component method as controller" do
+    notice = build_notice
+    notice.component = 'users_controller'
+
+    expect(notice.controller).to eq 'users_controller'
+  end
+
+  it "aliases component= method as controller=" do
+    notice = build_notice
+    notice.controller = 'users_controller'
+
+    expect(notice.component).to eq 'users_controller'
+  end
+
   it "aliases the params as parameters" do
     expect(build_notice(parameters: {foo: 'foo'}).params).to eq({foo: 'foo'})
     expect(build_notice(params: {bar: 'bar'}).parameters).to eq({bar: 'bar'})
@@ -795,21 +809,25 @@ describe Honeybadger::Notice do
   context "adding tags" do
     context "directly" do
       it "converts String to tags Array" do
-        expect(build_notice(tags: ' foo  , bar, ,  $baz   ').tags).to eq(%w(foo bar baz))
+        expect(build_notice(tags: ' foo  , bar, ,  baz   ').tags).to eq(%w(foo bar baz))
       end
 
       it "accepts an Array" do
-        expect(build_notice(tags: [' foo  ', ' bar', ' ', '  $baz   ']).tags).to eq(%w(foo bar baz))
+        expect(build_notice(tags: [' foo  ', ' bar', ' ', '  baz   ']).tags).to eq(%w(foo bar baz))
+      end
+
+      it "accepts whitespace-delimited tags" do
+        expect(build_notice(tags: [' foo bar  baz']).tags).to eq(%w(foo bar baz))
       end
     end
 
     context "from context" do
       it "converts String to tags Array" do
-        expect(build_notice(context: { tags: ' foo  , , bar,  $baz   ' }).tags).to eq(%w(foo bar baz))
+        expect(build_notice(context: { tags: ' foo  , , bar,  baz   ' }).tags).to eq(%w(foo bar baz))
       end
 
       it "accepts an Array" do
-        expect(build_notice(tags: [' foo  ', ' bar', ' ', '  $baz   ']).tags).to eq(%w(foo bar baz))
+        expect(build_notice(tags: [' foo  ', ' bar', ' ', '  baz   ']).tags).to eq(%w(foo bar baz))
       end
     end
 
@@ -821,6 +839,10 @@ describe Honeybadger::Notice do
 
     it "converts nil to empty Array" do
       expect(build_notice(tags: nil).tags).to eq([])
+    end
+
+    it "allows non-word characters in tags while stripping whitespace" do
+      expect(build_notice(tags: 'word,  with_underscore ,with space, with-dash,with$special*char').tags).to eq(%w(word with_underscore with space with-dash with$special*char))
     end
   end